First time check in
These are map and music files for both Angel Island and Siege Perilous.
The server doesn't care about the music files, but both client and server need to be using the same map files to avoid 'stuck' spots.
The map has a few patches, and we added an orc fort. The music is all the original Ultima Online .MIDI files converted to high quality MP3.